backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pyjwt
passlib[bcrypt]
websockets
python-multipart
